[Refactor] #146 OrderListItemViewHolder를 독립적인 클래스로 변경
