+++ presentation/src/main/java/com/woowahan/banchan/ui/order/OrderListAdapter.kt
@@ -1,10 +1,8 @@
 package com.woowahan.banchan.ui.order
 
-import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
-import com.woowahan.banchan.databinding.ItemOrderListBinding
 import com.woowahan.domain.model.OrderModel
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -14,7 +12,7 @@ import timber.log.Timber
 
 class OrderListAdapter(
     private val itemClickListener: (OrderModel) -> Unit
-): RecyclerView.Adapter<OrderListAdapter.OrderListItemViewHolder>() {
+): RecyclerView.Adapter<OrderListItemViewHolder>() {
 
     private val deliveryStatePayload = "delivery_state_payload"
 
@@ -59,37 +57,6 @@ class OrderListAdapter(
 
     override fun getItemCount(): Int {
         return orders.size
-    }
-
-    class OrderListItemViewHolder(
-        private val binding: ItemOrderListBinding,
-        val itemClickListener: (OrderModel) -> Unit
-    ): RecyclerView.ViewHolder(binding.root){
-        companion object {
-            fun from(
-                parent: ViewGroup,
-                itemClickListener: (OrderModel) -> Unit
-            ): OrderListItemViewHolder = OrderListItemViewHolder(
-                ItemOrderListBinding.inflate(LayoutInflater.from(parent.context), parent, false),
-                itemClickListener
-            )
-        }
-
-        fun bind(item: OrderModel){
-            item.items.firstOrNull()?.let {
-                binding.imageUrl = it.imageUrl
-                binding.title = it.title
-            }
-            binding.price = item.totalPrice + item.deliveryFee
-            bindDeliveryState(item.deliveryState)
-            binding.itemSize = item.items.size
-            binding.order = item
-            binding.holder = this
-        }
-
-        fun bindDeliveryState(deliveryState: Boolean){
-            binding.deliveryItemState = deliveryState
-        }
     }
 
     class OrderListDiffUtilCallback(
